// apps/admin-ui/src/types/Instance.d.ts
export interface GroupedHardware {
  cpu: Cpu[];
  display: Display;
  memory: Memory;
  multimedia: Multimedia;
  network: GroupedHardwareNetwork[] | string;
  pci: Pci[];
  scsi: Scsi[];
  storage: Storage[] | string;
  system: System;
  usb: Usb[];
}

type HardwareDescription = [
  attribute: string,
  title: string,
  value: number | null,
];

interface InstanceAlert {
  type: string;
  summary: string;
  severity: "warning" | "danger" | "info";
}

interface InstanceUpgrades {
  regular: number;
  security: number;
}
// ...
interface UbuntuProInfo {
  attached: boolean;
  expires: string | null;
  services: UbuntuProService[];
  techSupportLevel?: string;
  effective?: string;
  contract?: UbuntuProContract;
  account?: UbuntuProAccount;
  result: "success";
}

interface FailedUbuntuProInfo {
  errors: {
    message: string;
    message_code: string;
    service: string | null;
    type: string;
  }[];
  result: "failure";
}

interface DistributionInfo {
  code_name: string | null;
  description: string;
  distributor: string;
  release: string;
}

export interface InstanceWithoutRelation extends Record<string, unknown> {
  access_group: AccessGroup["name"];
  cloud_init: {
    availability_zone?: string | null;
  };
  comment: string;
  distribution: string | null;
  distribution_info: DistributionInfo | null;
  employee_id: number | null;
  hostname: string;
  id: number;
  is_default_child: boolean | null;
  is_wsl_instance: boolean;
  last_ping_time: string | null;
  tags: string[];
  title: string;
  ubuntu_pro_info: UbuntuProInfo | FailedUbuntuProInfo | null;
  annotations?: Record<string, string>;
  grouped_hardware?: GroupedHardware;
  alerts?: InstanceAlert[];
  upgrades?: InstanceUpgrades;
  archived: boolean;
}

interface WithRelation<Type extends InstanceWithoutRelation> extends Type {
  children: InstanceWithoutRelation[];
  parent: InstanceWithoutRelation | null;
}

export type Instance = WithRelation<InstanceWithoutRelation>;